mod_netv: Improve end-of-program reliability
When a program terminates, don't just close filehandles.  Instead,
mark the lua entry as available for reuse, and just sit on the
file handles.  This lets them get flushed as-needed.

Also, when we hit end-of-file, set the HTTP status as 204 and return
an empty HTTP response.  This allows the client to know that there's
no more data.